Add POC implementation of URL creation
This proof-of-concept retrieves the web page and some of its details to
generate the descriptive URL. Simple endpoints to get and post new URLs are
added, as well as a simple endpoint to redirect to the target URL.
functionality is added too.

The database is volatile (only kept in memory) for now.