Retrieve web page asynchronously
So far the web page from which to generate the expanded URL is retrieved
synchronously, but this is problematic because the worker thread will be
stuck waiting for data to arrive from the network.

We convert the web page retrieval to use the async framework, so the
worker thread can be used to process other requests while we wait for
data to come from the network.

Part of #19.