Use custom parsing of HTML
Replace webpage-rs with kuchiki for parsing and traversing the HTML to
look for useful information to expand the URL.

webpage-rs does much more than what we need: it will retrieve the page
and parse it completely, storing the whole body and even
the text-only extracted body. These are OK for a simple script (or even
some user facing app) but it's too wasteful for a server that should
eventually handle thousands of retrievals per second.

There is also some shuffling around with the struct used to store the
retrieved information but for now we should not focus on that and just
care about the expanded URL to be minimally informative.

Fixes #18.