Persist URLs to SQLite
So far all the state was only kept in memory, but of course we need to
keep the data even after any type of restart/crash.

This commit persists state through a SQLite local database. This is just
a temporary solution while prototyping, the final solution should use a
distributed database (probably a key-value store).

Fixes #16.

Signed-off-by: Leandro Lucarella <luca@llucax.com>