Save the complete expanded URL to the DB
We were saving only the expanded path component in the DB, not the
complete URL with a complete schema including the "view" host and URL
ID. This is to make it easier to port a DB to a different "view" host,
but it should actually be a very rare case and it should be easy enough
to batch-convert the DB if that's ever needed. This would avoid doing
any processing to build the noclick_url when retrieving the record,
which should be a more common case than changing the "view" host.

All that said, it's worth mentioning that the noclick_url is being saved
but not really used. It could be not saved at all, but for now we keep
it for debugging purposes. It might be a good thing for privacy to not
save it at all.

Signed-off-by: Leandro Lucarella <luca@llucax.com>