Fix host-based guard for /
The guard was using req.uri.host, which is only set when the request
comes via HTTP/2 (see https://github.com/actix/actix-web/issues/1050).
We use the guard::Host which takes differences between HTTP/1 and HTTP/2
into account. The path is still guarded so the whole /* is not handled
by the webapp service (we only want to handle /).

Signed-off-by: Leandro Lucarella <luca@llucax.com>